AJustes facturación

+++ database/migrations/2021_06_08_114819_create_invoice_items_table.php
@@ -19,7 +19,7 @@ class CreateInvoiceItemsTable extends Migration
             $table->foreign('invoice_id')->references('id')->on('invoices');
             $table->text('item')->nullable();
             $table->double('price')->nullable();
-            $table->double('quantity')->nullable();
+            $table->integer('quantity')->nullable();
             $table->double('value_final')->nullable();
             $table->timestamps();
         });
